requirements.txt
discord.py
riotwatcher
python-dotenv
tabulate
beautifulsoup4
requests
